# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * ComptaSimple TPE API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Client {
  id: number;
  name: string;
  /** @nullable */
  email?: string | null;
  /** @nullable */
  siret?: string | null;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  address?: string | null;
  /** @nullable */
  city?: string | null;
  /** @nullable */
  postalCode?: string | null;
  /** @nullable */
  legalForm?: string | null;
  /** @nullable */
  website?: string | null;
  /** @nullable */
  iban?: string | null;
  createdAt: string;
}

export interface ClientInput {
  /** @minLength 1 */
  name: string;
  email?: string;
  siret?: string;
  phone?: string;
  address?: string;
  city?: string;
  postalCode?: string;
  legalForm?: string;
  website?: string;
  iban?: string;
}

export interface ClientUpdate {
  /** @minLength 1 */
  name?: string;
  email?: string;
  siret?: string;
  phone?: string;
  address?: string;
  city?: string;
  postalCode?: string;
  legalForm?: string;
  website?: string;
  iban?: string;
}

export interface Invoice {
  id: number;
  invoiceId: string;
  clientName: string;
  date: string;
  amountHT: string;
  vatRate: string;
  status: string;
  createdAt: string;
}

export interface InvoiceInput {
  invoiceId: string;
  clientName: string;
  date: string;
  amountHT: string;
  vatRate: string;
  status?: string;
}

export interface InvoiceUpdate {
  clientName?: string;
  date?: string;
  amountHT?: string;
  vatRate?: string;
  status?: string;
}

export interface DevisLine {
  id: string;
  description: string;
  quantity: number;
  unitPrice: number;
  unit: string;
  vatRate: number;
}

export interface Devis {
  id: number;
  devisId: string;
  clientName: string;
  subject: string;
  date: string;
  validUntil: string;
  lines: DevisLine[];
  /** @nullable */
  notes?: string | null;
  /** @nullable */
  paymentTerms?: string | null;
  status: string;
  createdAt: string;
}

export interface DevisInput {
  devisId: string;
  clientName: string;
  subject: string;
  date: string;
  validUntil: string;
  lines: DevisLine[];
  notes?: string;
  paymentTerms?: string;
  status?: string;
}

export interface DevisUpdate {
  clientName?: string;
  subject?: string;
  date?: string;
  validUntil?: string;
  lines?: DevisLine[];
  notes?: string;
  paymentTerms?: string;
  status?: string;
}

export interface Expense {
  id: number;
  expenseId: string;
  supplier: string;
  date: string;
  amountHT: string;
  vatRate: string;
  category: string;
  createdAt: string;
}

export interface ExpenseInput {
  expenseId: string;
  supplier: string;
  date: string;
  amountHT: string;
  vatRate: string;
  category?: string;
}

export interface Transaction {
  id: number;
  txId: string;
  date: string;
  label: string;
  amount: string;
  type: string;
  /** @nullable */
  category?: string | null;
  /** @nullable */
  vatRate?: string | null;
  /** @nullable */
  documentName?: string | null;
  createdAt: string;
}

export interface TransactionInput {
  txId: string;
  date: string;
  label: string;
  amount: string;
  type: string;
  category?: string;
  vatRate?: string;
  documentName?: string;
}

export interface BulkTransactionInput {
  transactions: TransactionInput[];
}

export interface TransactionUpdate {
  category?: string;
  /** @nullable */
  vatRate?: string | null;
  /** @nullable */
  documentName?: string | null;
}

export interface BankAccount {
  id: number;
  name: string;
  iban: string;
  connected: boolean;
  updatedAt?: string;
}

export interface BankAccountInput {
  name: string;
  iban: string;
  connected: boolean;
}

export interface Company {
  id: number;
  name: string;
  legalForm: string;
  siret: string;
  vatNumber: string;
  address: string;
  postalCode: string;
  city: string;
  phone: string;
  email: string;
  website: string;
  iban: string;
  logo?: string;
  updatedAt?: string;
}

export interface CompanyInput {
  name?: string;
  legalForm?: string;
  siret?: string;
  vatNumber?: string;
  address?: string;
  postalCode?: string;
  city?: string;
  phone?: string;
  email?: string;
  website?: string;
  iban?: string;
  logo?: string;
}

export interface AiChatMessage {
  role: string;
  content: string;
}

export type AiChatInputContext = { [key: string]: unknown };

export interface AiChatInput {
  messages: AiChatMessage[];
  context?: AiChatInputContext;
}